Fix not working slack.
As reported in #23 and #27 slack is currently broken in the atomic
implementation.

Both #23 and #27 propose valid solutions, and I actually arrived
at #23 when trying to solve this independently, but I find #27
slightly more readable/understandable, so going with that.

Also, #23 seems to expose another bug in andres-erbsen/clock, with
sleeping with negative time, so we should probably consider switching
to benbjohnson/clock - maybe in a follow up.